+++ src/products/products.service.ts
@@ -7,6 +7,7 @@ import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { Product } from './entities/product.entity';
 import validate from './validate/validate-product';
+import * as fs from 'fs';
 
 @Injectable()
 export class ProductsService {
@@ -17,7 +18,7 @@ export class ProductsService {
     private cateRepository: Repository<Category>
   ){}
 
-  async create(createProductDto: CreateProductDto) {
+  async create(createProductDto: CreateProductDto, file: Express.Multer.File) {
     const category = await this.cateRepository.findOne(+createProductDto.categoryId);
     if(!category) {
       throw new HttpException({
@@ -26,12 +27,15 @@ export class ProductsService {
         redirect: '/admin/products/add'
       }, HttpStatus.BAD_REQUEST);
     }
-    const isValid = await validate(createProductDto, this.productRepository);
+    const isValid = await validate(createProductDto, this.productRepository, file);
     if(isValid === true) {
+      const arr =  file.path.split('/');
+      arr.splice(0, 1);
       createProductDto.slug = convertToSlug(createProductDto.name);
       
       const newCate = await this.productRepository.create({
         ...createProductDto,
+        image: arr.join('/'),
         category
       });
       return this.productRepository.save(newCate);
@@ -53,7 +57,10 @@ export class ProductsService {
 
   async findAll(page) {
     return await this.productRepository.find({
-      // relations: ['products'],
+      relations: ['category'],
+      order: {
+        'id': 'DESC'
+      },
       skip: page * (+process.env.PAGE_SIZE)  || 0,
       take: (+process.env.PAGE_SIZE) 
     });
@@ -64,6 +71,9 @@ export class ProductsService {
       withDeleted: true,
       where: {
         deleted_at: Not(IsNull())
+      },
+      order: {
+        'id': 'DESC'
       },
       skip: page * (+process.env.PAGE_SIZE)  || 0,
       take: (+process.env.PAGE_SIZE),
@@ -88,7 +98,7 @@ export class ProductsService {
   }
 
   async findOneTrash(id: number) {
-    return await this.productRepository.find({
+    return await this.productRepository.findOne({
       withDeleted: true,
       where: {
         id
@@ -105,25 +115,46 @@ export class ProductsService {
     });
   }
 
-  async update(id: number, updateProductDto: UpdateProductDto) {
+  async update(id: number, updateProductDto: UpdateProductDto, file: Express.Multer.File) {
     const product = await this.productRepository.findOne(id);
     
     if(!product) {
       throw new HttpException({
         status: HttpStatus.NOT_FOUND,
-        error: ['Loại sản phẩm không tồn tại!'],
+        error: ['Sản phẩm không tồn tại!'],
         redirect: `/admin/products`
       }, HttpStatus.NOT_FOUND);
     }
+
+    const category = await this.cateRepository.findOne(+updateProductDto.categoryId);
+    if(!category) {
+      throw new HttpException({
+        status: HttpStatus.BAD_REQUEST,
+        error: 'Loại sản phẩm không tồn tại!',
+        redirect: `/admin/products/update/${product.slug}`
+      }, HttpStatus.BAD_REQUEST);
+    }
     
-    const isValid = await validate(updateProductDto, this.productRepository, id);
+    const isValid = await validate(updateProductDto, this.productRepository, file,  id);
     if(isValid === true) {
-      updateProductDto.image = updateProductDto.image || product.image;
+      let path = '';
+      if(file) {
+        const arr =  file.path?.split('/');
+        arr.splice(0, 1);
+        path = arr.join('/');
+        fs.unlinkSync('./public/'+product.image);
+      }
+      else {
+        path = product.image;
+      }
+      
       updateProductDto.slug = convertToSlug(updateProductDto.name);
       
       const newProduct = await this.productRepository.preload({
         id: +id,
-        ...updateProductDto
+        ...updateProductDto,
+        image: path,
+        category
       });
       
       return this.productRepository.save(newProduct);
@@ -145,7 +176,19 @@ export class ProductsService {
     return await this.productRepository.restore(id);
   }
 
-  remove(id: number) {
-    return this.productRepository.delete(id);
+  async remove(id: number) {
+    const product = await this.findOneTrash(+id);
+
+    if(!product) {
+      throw new HttpException({
+        status: HttpStatus.BAD_REQUEST,
+        error: 'Sản phẩm không tồn tại!',
+        redirect: `/admin/products`
+      }, HttpStatus.BAD_REQUEST);
+    }
+    else {
+      fs.unlinkSync('./public/'+product.image);
+    }
+    return await this.productRepository.delete(id);
   }
 }
